+++ www/src/js/types/export.ts
@@ -3,6 +3,8 @@ import { ModuleCode, Semester } from 'types/modules';
 import { Mode } from 'types/settings';
 import { ColorMapping, ThemeState } from 'types/reducers';
 
+/* eslint-disable import/prefer-default-export */
+
 export type ExportData = {
   readonly semester: Semester;
   readonly timetable: SemTimetableConfig;
