+++ www/src/js/types/json2mq.d.ts
@@ -1,8 +1,7 @@
 declare module 'json2mq' {
   export interface QueryObject {
     [property: string]: string | number | boolean;
-    [index: number]: QueryObject;
   }
 
-  export default function json2mq(query: QueryObject): string;
+  export default function json2mq(query: QueryObject | QueryObject[]): string;
 }
