+++ www/src/js/views/planner/PlannerModule.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import Downshift from 'downshift';
 import { Draggable } from 'react-beautiful-dnd';
 import { Link } from 'react-router-dom';
 import { format } from 'date-fns';
@@ -10,9 +9,11 @@ import { Conflict } from 'types/views';
 import config from 'config';
 import { examDateToDate, renderMCs } from 'utils/modules';
 import { conflictToText } from 'utils/planner';
-import { AlertTriangle, ChevronDown } from 'views/components/icons';
+import { AlertTriangle } from 'views/components/icons';
 import LinkModuleCodes from 'views/components/LinkModuleCodes';
 import { modulePage } from 'views/routes/paths';
+
+import ModuleMenu from './ModuleMenu';
 import styles from './PlannerModule.scss';
 
 type Props = Readonly<{
@@ -30,68 +31,6 @@ type Props = Readonly<{
   removeModule: (moduleCode: ModuleCode) => void;
   addCustomData: (moduleCode: ModuleCode) => void;
 }>;
-
-type MenuProps = {
-  readonly removeModule: () => void;
-  readonly editCustomData: () => void;
-};
-
-type MenuItem = {
-  label: string;
-  action: () => void;
-  className?: string;
-};
-
-const ModuleMenu = React.memo((props: MenuProps) => {
-  const menuItems: MenuItem[] = [
-    { label: 'Edit MC and Title', action: props.editCustomData },
-    { label: 'Remove', action: props.removeModule, className: 'dropdown-item-danger' },
-  ];
-
-  return (
-    <Downshift
-      onSelect={(item) => {
-        menuItems.forEach(({ label, action }) => {
-          if (item === label) {
-            action();
-          }
-        });
-      }}
-    >
-      {({ getItemProps, getMenuProps, highlightedIndex, isOpen, toggleMenu }) => (
-        <div className={styles.menuBtn}>
-          <button
-            className={classnames('btn close')}
-            type="button"
-            onClick={() => toggleMenu()}
-            data-toggle="dropdown"
-            aria-haspopup="true"
-            aria-expanded={isOpen}
-          >
-            <ChevronDown />
-          </button>
-          <div
-            className={classnames(styles.menu, 'dropdown-menu', { show: isOpen })}
-            {...getMenuProps()}
-          >
-            {menuItems.map(({ label, className }, itemIndex) => (
-              <button
-                type="button"
-                key={label}
-                className={classnames('dropdown-item', className, {
-                  'dropdown-selected': highlightedIndex === itemIndex,
-                })}
-                {...getItemProps({ item: label })}
-              >
-                {label}
-              </button>
-            ))}
-          </div>
-        </div>
-      )}
-    </Downshift>
-  );
-});
 
 /**
  * Component for a single module on the planner
