+++ www/src/js/views/venues/VenueList.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
 import classnames from 'classnames';
 import { groupBy, toPairs, sortBy } from 'lodash';
-import { Link } from 'react-router-dom';
+import { Link, LinkProps } from 'react-router-dom';
 
+import { Omit } from 'types/utils';
 import { Venue } from 'types/venues';
 import { venuePage } from 'views/routes/paths';
 
@@ -11,7 +12,7 @@ import styles from './VenueList.scss';
 type Props = {
   venues: Venue[];
   selectedVenue?: Venue | null;
-  linkProps?: { [key: string]: any };
+  linkProps?: Omit<LinkProps, 'to'>;
 };
 
 export default function VenueList(props: Props) {
